requirements.txt
fastapi
uvicorn
requests
xmltodict
python-dotenv
openapi-spec-validator
